api/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-docx
pypdf
